refactor: move repository interfaces to domain layer
