+++ apptoolkit/src/main/java/com/d4rk/android/libs/apptoolkit/app/issuereporter/domain/repository/IssueReporterRepository.kt
@@ -1,4 +1,4 @@
-package com.d4rk.android.libs.apptoolkit.app.issuereporter.data
+package com.d4rk.android.libs.apptoolkit.app.issuereporter.domain.repository
 
 import com.d4rk.android.libs.apptoolkit.app.issuereporter.domain.model.IssueReportResult
 import com.d4rk.android.libs.apptoolkit.app.issuereporter.domain.model.Report
